requirements.txt
# Web Framework
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0

# Database - PostgreSQL Support
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Data Processing & AI/ML
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
click==8.1.7

# Werkzeug (Flask dependency)
Werkzeug>=3.0.0

# Forecasting
statsmodels==0.14.1
